Father, forgive me, for I am adding inline types.
