+++ robyn/__init__.py
@@ -12,6 +12,7 @@ from robyn.dev_event_handler import EventHandler
 from robyn.log_colors import Colors
 from robyn.processpool import spawn_process
 from robyn.responses import jsonify, static_file
+from typing import Callable, Optional
 
 from robyn.robyn import SocketHeld
 from robyn.router import MiddlewareRouter, Router, WebSocketRouter
@@ -23,7 +24,7 @@ logger = logging.getLogger(__name__)
 class Robyn:
     """This is the python wrapper for the Robyn binaries."""
 
-    def __init__(self, file_object):
+    def __init__(self, file_object: str) -> None:
         directory_path = os.path.dirname(os.path.abspath(file_object))
         self.file_path = file_object
         self.directory_path = directory_path
@@ -54,7 +55,7 @@ class Robyn:
         """
         self.router.add_route(route_type, endpoint, handler, const)
 
-    def before_request(self, endpoint):
+    def before_request(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.before_request decorator to add a get route
 
@@ -63,7 +64,7 @@ class Robyn:
 
         return self.middleware_router.add_before_request(endpoint)
 
-    def after_request(self, endpoint):
+    def after_request(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.after_request decorator to add a get route
 
@@ -73,17 +74,17 @@ class Robyn:
         return self.middleware_router.add_after_request(endpoint)
 
     def add_directory(
-        self, route, directory_path, index_file=None, show_files_listing=False
+        self, route: str, directory_path: str, index_file: Optional[str] = None, show_files_listing: bool = False
     ):
         self.directories.append((route, directory_path, index_file, show_files_listing))
 
-    def add_header(self, key, value):
+    def add_header(self, key: str, value: str) -> None:
         self.headers.append((key, value))
 
-    def add_web_socket(self, endpoint, ws):
+    def add_web_socket(self, endpoint: str, ws: WS) -> None:
         self.web_socket_router.add_route(endpoint, ws)
 
-    def _add_event_handler(self, event_type: str, handler):
+    def _add_event_handler(self, event_type: str, handler) -> None:
         logger.debug(f"Add event {event_type} handler")
         if event_type not in {Events.STARTUP, Events.SHUTDOWN}:
             return
@@ -91,13 +92,13 @@ class Robyn:
         is_async = asyncio.iscoroutinefunction(handler)
         self.event_handlers[event_type] = (handler, is_async)
 
-    def startup_handler(self, handler):
+    def startup_handler(self, handler: Callable) -> None:
         self._add_event_handler(Events.STARTUP, handler)
 
-    def shutdown_handler(self, handler):
+    def shutdown_handler(self, handler: Callable) -> None:
         self._add_event_handler(Events.SHUTDOWN, handler)
 
-    def start(self, url="127.0.0.1", port=5000):
+    def start(self, url: str = "127.0.0.1", port: int = 5000):
         """
         Starts the server
 
@@ -157,7 +158,6 @@ class Robyn:
                 logger.info(f"{Colors.BOLD}{Colors.OKGREEN} Terminating server!! {Colors.ENDC}")
                 for process in process_pool:
                     process.kill()
-
         else:
             event_handler = EventHandler(self.file_path)
             event_handler.start_server_first_time()
@@ -174,7 +174,7 @@ class Robyn:
                 observer.stop()
                 observer.join()
 
-    def get(self, endpoint, const=False):
+    def get(self, endpoint: str, const: bool = False) -> Callable[..., None]:
         """
         The @app.get decorator to add a get route
 
@@ -186,7 +186,7 @@ class Robyn:
 
         return inner
 
-    def post(self, endpoint):
+    def post(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.post decorator to add a get route
 
@@ -198,7 +198,7 @@ class Robyn:
 
         return inner
 
-    def put(self, endpoint):
+    def put(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.put decorator to add a get route
 
@@ -209,7 +209,7 @@ class Robyn:
 
         return inner
 
-    def delete(self, endpoint):
+    def delete(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.delete decorator to add a get route
 
@@ -221,7 +221,7 @@ class Robyn:
 
         return inner
 
-    def patch(self, endpoint):
+    def patch(self, endpoint: str) -> Callable[..., None]:
         """
         [The @app.patch decorator to add a get route]
 
@@ -233,7 +233,7 @@ class Robyn:
 
         return inner
 
-    def head(self, endpoint):
+    def head(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.head decorator to add a get route
 
@@ -245,7 +245,7 @@ class Robyn:
 
         return inner
 
-    def options(self, endpoint):
+    def options(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.options decorator to add a get route
 
@@ -257,7 +257,7 @@ class Robyn:
 
         return inner
 
-    def connect(self, endpoint):
+    def connect(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.connect decorator to add a get route
 
@@ -269,7 +269,7 @@ class Robyn:
 
         return inner
 
-    def trace(self, endpoint):
+    def trace(self, endpoint: str) -> Callable[..., None]:
         """
         The @app.trace decorator to add a get route
