+++ robyn/events.py
@@ -1,3 +1,6 @@
-class Events:
+from enum import Enum
+
+
+class Events(Enum):
     STARTUP = "startup"
     SHUTDOWN = "shutdown"
