--- robyn/log_colors.pyi
@@ -1,10 +1,0 @@
-class Colors:
-    HEADER = ...
-    OKBLUE = ...
-    OKCYAN = ...
-    OKGREEN = ...
-    WARNING = ...
-    FAIL = ...
-    ENDC = ...
-    BOLD = ...
-    UNDERLINE = ...
