--- robyn/responses.pyi
@@ -1,19 +1,0 @@
-from typing import Any
-
-def static_file(file_path: str) -> dict[str, Any]:
-    """
-    This function will help in serving a static_file
-
-    :param file_path str: file path to serve as a response
-    """
-
-    pass
-
-def jsonify(input_dict: object) -> str:
-    """
-    This function serializes input dict to a json string
-
-    :param input_dict object: response of the function
-    """
-
-    pass
